backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.11.0
python-dotenv==1.0.1
sentence-transformers==3.3.1
httpx==0.28.1
pydantic==2.10.4
